Review Changes: Remove useless method + renamed hooks
