+++ app/templates/layers/page.tsx
@@ -4,15 +4,14 @@ import {useCallback, useState} from 'react';
 import {LayerMode} from '../../../src/app/types/layerMode.types';
 import {Player} from '@remotion/player';
 import {Form, Input} from '../../../src/app/forms/input';
-import {useInputChange} from '../../../src/app/hooks/onInputChange';
+import {useInputChange} from '../../../src/app/hooks/useInputChange';
 import LayerByMode from '../../../src/app/LayerByMode';
 import {SelectInput} from '../../../src/app/forms/selectInput';
 import {ColorInput} from '../../../src/app/forms/colorInput';
-import {useSelectChange} from '../../../src/app/hooks/onSelectChange';
 
 export default function LayersPage() {
 	const [copied, setCopied] = useState(false);
-	const [mode, setMode] = useSelectChange<LayerMode>('one');
+	const [mode, setMode] = useInputChange<LayerMode>('one');
 	const [title, setTitle] = useInputChange<string>('Shortvid.io 🎬');
 	const [sponsor, setSponsor] = useInputChange<string>(
 		'https://secure.meetupstatic.com/photos/event/2/8/7/a/600_464230362.jpeg'
