+++ app/templates/talk/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import {Player} from '@remotion/player';
-import {useInputChange} from '../../../src/app/hooks/onInputChange';
+import {useInputChange} from '../../../src/app/hooks/useInputChange';
 import {Form, Input} from '../../../src/app/forms/input';
 import {Code} from '../../../src/app/Code';
 import {Talk} from '../../../remotion/compositions/templates/talk/Talk';
